[add] content switching logic to app
