+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "agroeasy",
-  "version": "1.1.0",
+  "version": "1.1.2",
   "description": "",
   "main": "server.js",
   "scripts": {
