+++ src/components/app/components/Footer.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Layout } from 'antd';
 
-import { FOOTER_DATA, FOOTER_CLASSNAMES } from './constants';
+import { FOOTER_DATA, FOOTER_CLASSNAMES } from '../constants';
 
 const { Footer } = Layout;
 const { CONTAINER, DYNAMIC_FOOTER, ROW } = FOOTER_CLASSNAMES;
@@ -9,11 +9,7 @@ const { CONTAINER, DYNAMIC_FOOTER, ROW } = FOOTER_CLASSNAMES;
 const footers = FOOTER_DATA.map(footer => (
     <div className={DYNAMIC_FOOTER} key={footer.key}>
         <h3>{footer.header}</h3>
-        {
-            footer.items.map(item => (
-                <div key={item.title}>{item.title}</div>
-            ))
-        }
+        {footer.items.map(item => <div key={item.title}>{item.title}</div>)}
     </div>
 ));
