+++ src/components/app/components/index.js
@@ -1,0 +1,4 @@
+import App from './App';
+import './styles.css';
+
+export { App };
