+++ src/components/app/components/styles.css
@@ -1,0 +1,4 @@
+.footer-container {
+    background-color: #f0f2f5;
+    text-align: center;
+}
