+++ src/components/home/components/ContactUs.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button, Modal } from 'antd';
 
-import { CONTACT_US } from './constants';
+import { CONTACT_US } from '../constants';
 import FormContent from './FormContent';
 
 const { BACK, CANCEL, PRIMARY, SMALL, SUBMIT, TITLE } = CONTACT_US;
