+++ src/components/home/components/Home.js
@@ -3,11 +3,9 @@ import { Col, message, Row } from 'antd';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
-import AppFoot from './Footer';
-import Navibar from './Navibar';
 import Search from './Search';
 import { components } from '../../productList';
-import { HOME_STRINGS } from './constants';
+import { HOME_STRINGS } from '../constants';
 
 const { ProductList } = components;
 const { BG_IMG, SEARCH, SM_IMG, WELCOME_TEXT_1, WELCOME_TEXT_2, WELCOME_TEXT_3 } = HOME_STRINGS;
@@ -19,18 +17,17 @@ class Home extends React.Component {
         const { data } = signupState;
         return (
             <div>
-                <Navibar />
                 <div className={BG_IMG}>
                     <div >
-                       
+
                         {
-                            data.success !== undefined &&                           
+                            data.success !== undefined &&
                             <span>
                                 {
                                     data.success ? message.success(data.message, 5) :
                                         message.error(data.message, 5)
-                                } 
-                            </span>                 
+                                }
+                            </span>
                         }
                     </div>
                     <Row>
@@ -45,7 +42,6 @@ class Home extends React.Component {
                     </Row>
                 </div>
                 <ProductList />
-                <AppFoot />
             </div>
         );
     }
