+++ src/components/home/components/styles.css
@@ -1,9 +1,4 @@
 /* footer.js */
-.footer-container {
-    background-color: #f0f2f5;
-    text-align: center;
-}
-
 .details {
     display: inline-block;
     list-style: none;
