+++ src/components/home/constants.js
@@ -1,0 +1,35 @@
+/* links of Navibar */
+export const HOME = "/";
+export const SUPPORT = "/about/";
+export const MARKET = "/market/";
+
+export const CONTACT_US = {
+    BACK: "back",
+    CANCEL: "Cancel",
+    EMAIL: "Email",
+    MESSAGE: "Message",
+    NAME: "Name",
+    PRIMARY: "primary",
+    SMALL: "small",
+    SUBMIT: "Submit",
+    TITLE: "Contact Us",
+};
+
+//strings from ListItems.js
+export const SEARCH = {
+    CENTER: "center",
+    FLEX: "flex",
+    LARGE: "large",
+    SEARCH_PRODUCTS: "search products...",
+};
+
+//strings from Home.js
+export const HOME_STRINGS = {
+    BG_IMG: 'bg-img',
+    SEARCH: "search",
+    SM_IMG: "sm-img",
+    WELCOME_TEXT_1: "Welcome Guest",
+    WELCOME_TEXT_2: "Make us your distributor and we will tell the world about your business.",
+    WELCOME_TEXT_3: `Dear customer, we make sure you get a discount anytime you
+        patronize us. Give us a try.`,
+};
