+++ src/components/router/components/Routes.js
@@ -1,0 +1,26 @@
+import React from 'react';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
+
+import app from '../../app';
+import home from '../../home';
+import NoMatch from './NoMatch';
+import userProfile from '../../userProfile';
+import { PATHS } from '../constants';
+
+const { HOME, MARKET, PROFILE } = PATHS;
+const { App } = app.components;
+
+const Routes = () => (
+    <BrowserRouter>
+        <App>
+            <Switch>
+                <Route path={HOME} exact strict component={home.components.Home} />
+                <Route path={MARKET} exact strict component={home.components.Home} />
+                <Route path={PROFILE} exact strict component={userProfile.components.UserProfile} />
+                <Route component={NoMatch} />
+            </Switch>
+        </App>
+    </BrowserRouter>
+);
+
+export default Routes;
