+++ src/components/router/components/index.js
@@ -1,0 +1,3 @@
+import Routes from './Routes';
+
+export { Routes };
