+++ src/components/router/constants.js
@@ -1,0 +1,9 @@
+/** @constant */
+export const PATHS = {
+    CONTACT: '/contact',
+    HOME: '/',
+    MARKET: '/market',
+    PROFILE: '/profile',
+    SIGN_IN: '/signin',
+    SIGN_UP: '/signup',
+};
