+++ src/components/signin/components/Signin.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import SigninForm from './SigninForm';
-import { SIGNIN_STRINGS } from './constants';
+import { SIGNIN_STRINGS } from '../constants';
 
 const { PRIMARY, TITLE } = SIGNIN_STRINGS;
 
@@ -9,35 +9,35 @@ export default class Signin extends React.Component {
     state = {
         visible: false,
     };
-  
+
     showModal = () => {
         this.setState({ visible: true });
     }
-  
+
     handleCancel = () => {
         this.setState({ visible: false });
     }
-  
+
     handleCreate = () => {
         const form = this.formRef.props.form;
         form.validateFields(error => {
             if (error) {
                 return error;
             }
-  
+
             form.resetFields();
             this.setState({ visible: false });
         });
     }
-  
+
     saveFormRef = formRef => {
         this.formRef = formRef;
     }
-   
+
     render() {
         return (
             <div>
-                <span type={PRIMARY} onClick={this.showModal}>{TITLE}</span>
+                <div type={PRIMARY} onClick={this.showModal}>{TITLE}</div>
                 <SigninForm
                     wrappedComponentRef={this.saveFormRef}
                     visible={this.state.visible}
