--- src/components/userPage/index.js
@@ -1,3 +1,0 @@
-import components from './components';
-
-export  { components };
