+++ src/components/userProfile/components/SideMenu.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Menu, Icon } from 'antd';
 
-import { SIDE_MENU_ITEMS, USER_PAGE } from './constants';
+import { SIDE_MENU_ITEMS, USER_PAGE } from '../constants';
 
-const { 
+const {
     CLASSNAMES: { MENU },
     STRINGS: { DARK, INLINE },
 } = USER_PAGE;
@@ -15,7 +15,7 @@ export default class SideMenu extends React.Component {
                 className={MENU}
                 theme={DARK}
                 mode={INLINE}
-            >   
+            >
                 {
                     SIDE_MENU_ITEMS.map(sideMenuItem => (
                         <Menu.Item key={sideMenuItem.label}>
