+++ src/components/userProfile/components/UserInfo.js
@@ -1,16 +1,18 @@
 import React from 'react';
 import { Card, Divider } from 'antd';
 
-import { CONTACT, LOCATION, USER_PAGE } from './constants';
+import { CONTACT, LOCATION, USER_PAGE } from '../constants';
 
 const { Meta } = Card;
-const { CLASSNAMES: { CARD_META, DATA_TITLE, HEADER_TITLE, INFO_CARD },
-    TEXTS: { CONTACT_INFO_TEXT, LOCATION_INFO_TEXT } } = USER_PAGE;
+const {
+    CLASSNAMES: { CARD_META, DATA_TITLE, HEADER_TITLE, INFO_CARD },
+    TEXTS: { CONTACT_INFO_TEXT, LOCATION_INFO_TEXT },
+} = USER_PAGE;
 
 export default class UserInfo extends React.Component {
     render() {
         return (
-            <Card 
+            <Card
                 className={INFO_CARD}
                 bordered={false}
             >
@@ -19,27 +21,27 @@ export default class UserInfo extends React.Component {
                     description={
                         CONTACT.map(contact => (
                             <div key={contact.title} className={DATA_TITLE}>
-                                <b>{contact.title}</b>                                       
+                                <b>{contact.title}</b>
                                 <div>
                                     {contact.description}
                                 </div>
                             </div>
-                        ))  
+                        ))
                     }
                 />
                 <Divider />
-                <Meta  
+                <Meta
                     className={CARD_META}
                     title={<h4 className={HEADER_TITLE}>{LOCATION_INFO_TEXT}</h4>}
                     description={
                         LOCATION.map(location => (
                             <div key={location.title} className={DATA_TITLE}>
-                                <b>{location.title}</b>                                       
+                                <b>{location.title}</b>
                                 <div>
                                     {location.description}
                                 </div>
                             </div>
-                        ))  
+                        ))
                     }
                 />
             </Card>
