+++ src/components/userProfile/components/index.js
@@ -1,4 +1,4 @@
 import UserProfile from './UserProfile';
 import './styles.css';
 
-export default { UserProfile };
+export { UserProfile };
