+++ src/components/userProfile/components/styles.css
@@ -1,21 +1,3 @@
-/*
-Navbar Component
-*/
-.main-nav-div {
-    background-color: white;
-}
-
-.navbar-menu {
-    background: white;
-    float: right;
-    padding-top: 10px;
-}
-.nav-items {
-    color: black !important;
-    font-weight: 500;
-    margin-left: 54px;
-    margin-right: 54px;
-}
 .hi-user {
     color: black !important;
     margin-left: 54px;
@@ -36,19 +18,19 @@ Navbar Component
 }
 .data-title {
     margin-top: 22px;
-    margin-bottom: 19.2px;  
+    margin-bottom: 19.2px;
 }
 .data-title div {
      float: right;
      font-size: 12.5px;
-     text-align: right;   
+     text-align: right;
 }
 .data-title b {
     font-size: 13.5px;
 }
 .basic-info {
     margin-top: 15px;
-    padding-bottom: 17.4px; 
+    padding-bottom: 17.4px;
 }
 .basic-info b {
     font-size: 13.5px;
@@ -56,7 +38,7 @@ Navbar Component
 .basic-info div {
     float: right;
     font-size: 13px;
-    padding-right: 40px; 
+    padding-right: 40px;
 }
 .big-card {
     background-color: DodgerBlue;
@@ -67,10 +49,10 @@ Navbar Component
     font-size: 16px;
     font-family: 'Lucida Grande', 'Lucida Sans Unicode', Verdana, sans-serif;
     font-weight: 600;
-    padding-left: 37px; 
+    padding-left: 37px;
 }
 .inner-card {
-    padding-right: 40px; 
+    padding-right: 40px;
     padding-left: 40px;
 }
 /**UserProfile.js */
@@ -78,17 +60,16 @@ Navbar Component
     background: white;
 }
 .big-layout {
-    background: white;  
-    padding: 40px 0;
+    background: white;
     margin-top: 1%;
 }
 .sider {
     background: white;
-   
+
 }
 .menu {
-    height: 126%;
-    padding-top: 20px;
+    height: 80vh;
+    margin-top: 50px;
 }
 .sm-content {
     padding: 0 14px;
@@ -107,7 +88,7 @@ Navbar Component
 /** UserAvatar */
 .avatar {
     display: block;
-    margin-left: auto; 
+    margin-left: auto;
     margin-right: auto;
 }
 .avatar-card {
